landlim/requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
bootstrap4==0.1.0
dj-database-url==0.5.0
Django==3.2.6
django-appconf==1.0.4
django-bootstrap4==3.0.1
django-cleanup==5.2.0
django-crispy-forms==1.11.2
django-file-validator==0.0.7
django-filter==21.1
django-heroku==0.3.1
django-imagekit==4.0.2
django-leaflet==0.28.1
django-summernote==0.8.11.6
django-taggit==1.5.1
django-taggit-templatetags==0.2.5
django-templatetag-sugar==1.0
django-versatileimagefield==2.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-gis==0.17
geographiclib==1.50
gunicorn==20.1.0
Markdown==3.3.4
pilkit==2.0
Pillow==8.2.0
psycopg2==2.8.6
python-magic==0.4.24
pytz==2021.1
rules==3.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
whitenoise==5.3.0
numpy==1.22.1
reverse-geocode==1.4.1
scipy==1.7.3
geopy==2.2.0
pyokaka==1.0.0
PyJWT==1.7.1